+++ CHANGES.rst
@@ -10,6 +10,7 @@ Improvements:
  * Widgets: Check user permission before opening jitsi (TODO design: #2842).
  * Widgets: Add a contextual menu to refresh, open outside, remove and revoke the permission (#2834).
  * Settings: Add an option for disabling use of the integration manager (#2843).
+ * Jitsi: Display room name, user name and user avatar in the conference screen.
 
 Changes in 0.10.2 (2019-11-15)
 ===============================================